Modularize oversized bandwidth, version, and preflight modules
