+++ crates/core/src/version/tests/mod.rs
@@ -1,0 +1,14 @@
+use super::*;
+use crate::branding::Brand;
+
+pub(super) const ACL_FROM_LABEL: Option<CompiledFeature> = CompiledFeature::from_label("ACLs");
+pub(super) const UNKNOWN_FROM_LABEL: Option<CompiledFeature> =
+    CompiledFeature::from_label("unknown");
+
+mod compiled_features;
+mod compiled_features_display;
+mod compiled_features_iter;
+mod compiled_features_static;
+mod metadata;
+mod report;
+mod secluded_args;
